% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Calculate accuracy metrics of two vectors}
\usage{
accuracy(reference, estimate, by = NULL, noinfo = TRUE,
  dist.normal = TRUE, short = FALSE)
}
\arguments{
\item{reference}{a vector of reference values}

\item{estimate}{a vector of estimated values}

\item{by}{Optional grouping variable}

\item{noinfo}{Logical. Should the additional information on the calculations be displayed?}

\item{dist.normal}{Logical. Is the distribution normal? If TRUE t-test and Pearson correlation is calculated. If FALSE, Wilcoxon paired test and Spearman correlation is calculated.}

\item{short}{Logical. Calculate only a subset of summary statistics.}
}
\value{
summary statistics of the differences between \code{reference} and \code{estimate}.
}
\description{
This is a direct call to \code{\link[scatter]{calc.error}}. Calculate bias, RMSE, correlation etc of two vectors.
}
\references{
\url{https://github.com/ptompalski/scatter}
}
