% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracyFoster.R
\name{accuracyFoster}
\alias{accuracyFoster}
\title{Calculate bias, RMSE, corelation etc of two vectors.}
\usage{
accuracyFoster(reference, estimate, by = NULL, noinfo = TRUE,
  dist.normal = TRUE, short = FALSE)
}
\arguments{
\item{reference}{a vector of reference values}

\item{estimate}{a vector of estimated values}

\item{by}{Optional grouping variable}

\item{noinfo}{Logical. Should the additional information on the calculations be displayed?}

\item{dist.normal}{Logical. Is the distribution normal? If TRUE t-test and Pearson correlation is calculated. If FALSE, Wilcoxon paired test and Spearman corelation is calculated.}

\item{short}{Logical. Calculate only a subset of summary statistics.}
}
\value{
summary statistics of the differences between \code{reference} and \code{estimate}.
}
\description{
Calculates: corelation coeficient,
 bias (absolute and relative), RMSE (absolute and relative), and performes statistical test to check if the differences between \code{reference} and \code{estimate} are significant.
 Correlation coefficient as well as statistical test depend on the \code{dist.normal} value.
 Optionally statistics can be calculated for groups of data, specified with the \code{by} parameter.
}
