% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allan_surv.R
\name{Allan_surv}
\alias{Allan_surv}
\title{Tick Survival Data Intake function for Brian Allan's AA ticks}
\usage{
Allan_surv(path = "", life.stage)
}
\arguments{
\item{path}{File path to csv if not in working directory}

\item{life.stage}{"Nymph" or "Adult"}
}
\description{
This funciton reads the raw csv file containing tick survival data
from the BA's survival experiments (soil cores) and transforms
the data to be fed to JAGS
}
\examples{
Allan_surv("Nymph")
}
