% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_param.R
\name{beta_param}
\alias{beta_param}
\title{Beta Distribution}
\usage{
beta_param(vec)
}
\arguments{
\item{vec}{vector with values between 0 and 1}
}
\description{
This function takes a vector with data that is [0, 1]
Uses mean and variance of input vector to estimate
the shape parameters (alpha & beta) for the beta distribution
}
\examples{
beta_param(vec)
}
