% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R, R/accuracyFoster.R
\name{calc_error}
\alias{calc_error}
\title{Calculate bias, RMSE, corelation etc of two vectors.}
\usage{
calc_error(reference, estimate)

calc_error(reference, estimate)
}
\arguments{
\item{reference}{a vector of reference values}

\item{estimate}{a vector of estimated values}

\item{by}{Optional grouping variable}

\item{noinfo}{Logical. Should the additional information on the calculations be displayed?}

\item{dist.normal}{Logical. Is the distribution normal? If TRUE t-test and Pearson correlation is calculated. If FALSE, Wilcoxon paired test and Spearman corelation is calculated.}

\item{short}{Logical. Calculate only a subset of summary statistics.}

\item{reference}{a vector of reference values}

\item{estimate}{a vector of estimated values}

\item{by}{Optional grouping variable}

\item{noinfo}{Logical. Should the additional information on the calculations be displayed?}

\item{dist.normal}{Logical. Is the distribution normal? If TRUE t-test and Pearson correlation is calculated. If FALSE, Wilcoxon paired test and Spearman corelation is calculated.}

\item{short}{Logical. Calculate only a subset of summary statistics.}
}
\value{
summary statistics of the differences between \code{reference} and \code{estimate}.

summary statistics of the differences between \code{reference} and \code{estimate}.
}
\description{
Internal function used by \code{\link[foster]{acuracy}}

Internal function used by \code{\link[foster]{acuracy}}
}
