% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_cary.R
\name{ch_cary}
\alias{ch_cary}
\title{Capture History Matrix Function}
\usage{
ch_cary(grid, path = "")
}
\arguments{
\item{grid}{What grid do you want to create matrix for? One of Green Control, Henry Control, Tea Control, Green Experimental, Henry Experimental, Tea Experimental}

\item{path}{File path to csv, Default = "" and assumes csv is in working directory}
}
\description{
This function reads the raw csv file containing white footef mice capture data
}
\examples{
ch.cary("Green Control")
}
