% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_TickStatesHB.R
\name{ci_TickStatesHB}
\alias{ci_TickStatesHB}
\title{CI's for tick state estimates by life stage from an HB model (includes all sites)
from Cary Institute}
\usage{
ci_TickStatesHB(state, site, nsamp = 5000)
}
\arguments{
\item{state}{matrix of iterations from bayes sampling}

\item{site}{What site to create CI's? One of "Green", "Henry", or "Tea"}

\item{nsamp}{How many random samples to draw? Default = 5000}
}
\description{
CI's for tick state estimates by life stage from an HB model (includes all sites)
from Cary Institute
}
\examples{
ci_TickStatesHB(jags.out, "Tea")
}
