% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSplit.R
\name{dataSplit}
\alias{dataSplit}
\title{Split data into training and testing sets}
\usage{
dataSplit(x, type = "group holdout", p = 0.75, groups = min(5,
  length(x)), returnTrain = TRUE)
}
\arguments{
\item{x}{A vector used for splitting data}

\item{type}{Character. How should data be splitted? Valid values are "random holdout" , "group holdout" or "kfold"}

\item{p}{percentage of data that goes to training set (holdout) or to each fold (1/k)}

\item{groups}{For "group holdout" and when x is numeric, this is the number of breaks in the quantiles}

\item{returnTrain}{Logical indicating whether training data or testing data should be returned}
}
\description{
Returns the indices of \code{x} that should go to trainign or testing
}
\details{
Three types of splits are currently implemented. "random holdout" randomly select \code{p} percents of \code{x} for the training set. 'group holdout" first group \code{x} into \code{groups} quantiles and randomly samples within them. "kfold" creates k folds where p percent of the data is used for training in each fold.
}
