% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defaultTemporalSummary}
\alias{defaultTemporalSummary}
\title{Calculate median, IQR and Theil Sen slope}
\usage{
defaultTemporalSummary(x)
}
\arguments{
\item{x}{Vector of numeric values}
}
\value{
A list with median, IQR and slope
}
\description{
This function is usually called within \code{\link[foster]{temporalMetrics}}
}
