% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\title{Remove cells in the neighorhood of a boundary cell}
\usage{
edges(x, w, filename = "", ...)
}
\arguments{
\item{x}{A \code{Raster* object}}

\item{w}{Numeric. Size of the window around each cell. Must be an odd number.}

\item{filename}{Character (optional). Output filename including path to
directory and eventually extension}

\item{...}{Additional arguments passed to \code{\link[raster]{writeRaster}}}
}
\description{
Assigns NA value to all cells having a NA values within their \code{w x w}
neighorhood.
}
\seealso{
\code{\link[raster]{focal}}
}
