% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgesFoster.R
\name{edgesFoster}
\alias{edgesFoster}
\title{Remove edges of input Raster}
\usage{
edgesFoster(x, w, filename = "", edgeVal = NA, ...)
}
\arguments{
\item{x}{A \code{Raster* object}}

\item{w}{Numeric. Window size around each cell. If one of the cell if the neighborhood had a edgeVal value, then all the cells inside the neighborhood will be removed.}

\item{filename}{Character (optional). If the output is written to disk, full path to location including filename}

\item{edgeVal}{value of cells that are on the outer side of edges}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}}
}
\description{
Remove edges of input Raster
}
