% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNN.R
\name{findNN}
\alias{findNN}
\title{Train a k-NN model}
\usage{
findNN(x, y, inTrain = NULL, inTest = NULL, k = 1,
  method = "randomForest", ntree = 200, mtry = NULL, ...)
}
\arguments{
\item{x}{A matrix or dataframe of predictors variables X for reference observations.Row names of X are used as identification of reference observations.}

\item{y}{A matrix or dataframe of response variables Y for the reference observations. Row names of Y are used as identification of reference observations.}

\item{inTrain}{Optional numeric vector indicating which rows of x and y go to training.}

\item{inTest}{Optional numeric vector indicating which rows of x and y go to training. If left NULL, all row that are not in inTrain are used for validation}

\item{k}{Integer. Number of nearest neighbors}

\item{method}{Character. Which nearness metrics is used to computed the nearest neighbors. Default is \code{"randomForest"}. Other methods are listed in \code{\link[yaImpute]{yai}}}

\item{ntree}{Number of classification or regression trees drawn for each response variable. Default is 200}

\item{mtry}{Number of X variables picked randomly to split each node. Default is sqrt(number of X variables)}

\item{...}{Other arguments passed to \code{\link[raster]{writeRaster}}}
}
\value{
A list containing the following objects:
   \describe{
   \item{\code{model}}{A \code{yai} object, the trained k-NN model}
   \item{\code{preds}}{A data.frame with observed and predicted values of the testing set for each response variables}
   \item{\code{accuracy}}{A data.frame with accuracy metrics of the trained k-NN model}
   }
}
\description{
This function trains a k-NN model from response variables (Y) and predictors (X) at reference location using the package yaImpute (see \code{\link[yaImpute]{yai}}). Optionally, training and testing sets can be provided to return the accuracy of the trained k-NN model.
}
\details{
If inTest = NULL, all rows that are not in inTrain will be used for model testing. If inTrain = NULL, all rows that are not in inTest will be used for model training. If both inTrain and inTest are NULL, all rows of x and y will be used for training and no testing is performed.
}
\seealso{
\code{\link[yaImpute]{yai}}, \code{\link[yaImpute]{newtargets}}, \code{\link[yaImpute]{impute.yai}}, \code{\link[foster]{accuracy}}
}
