% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalFoster.R
\name{focalFoster}
\alias{focalFoster}
\title{Apply a filter to image}
\usage{
focalFoster(x, w, fun, filename = "", na.rm = FALSE, pad = FALSE,
  padValue = NA, NAonly = FALSE, ...)
}
\arguments{
\item{x}{A \code{Raster* object}}

\item{w}{Matrix of weights}

\item{fun}{function (optional)}

\item{filename}{Character (optional). If the output is written to disk, full path to location including filename}

\item{na.rm}{logical. If TRUE, NA are removed from computation}

\item{pad}{logical. IF TRUE}

\item{padValue}{numeric}

\item{NAonly}{logical. If TRUE only cel values that are NA are replaced with the computed focal values}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}}
}
\description{
Apply a filter to image
}
