% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFoster.R
\name{fosterVarImp}
\alias{fosterVarImp}
\title{Variable importance}
\usage{
fosterVarImp(x, ntop = 20, plot = T, ...)
}
\arguments{
\item{x}{an object of class yai}

\item{plot}{if FALSE, no plotting is done, but the scores are returned}

\item{...}{passed to \code{\link[graphics]{boxplot}} function}

\item{nTop}{the nTop most important variables are plotted}
}
\value{
a plot of variable importance
}
\description{
Variable importance
}
