% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{getSampleXY}
\alias{getSampleXY}
\title{Generate a minimum distance stratified random sample}
\usage{
getSampleXY(x, n, mindist = 500)
}
\arguments{
\item{x}{A raster layer used to generate random sample}

\item{n}{Sample size}

\item{mindist}{Minimum distance between samples (in meters)}
}
\value{
A \code{SpatialPointsDataFrame} object containing sample locations
}
\description{
Generate a minimum distance stratified random sample
}
