% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_states.R
\name{known_states}
\alias{known_states}
\title{Known states function}
\usage{
known_states(ch)
}
\arguments{
\item{ch}{Capture history matrix}
}
\description{
This function fills in the capture history matrix where individuals are known but not observed
}
\examples{
known_states(ch)
}
