% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit Transformation}
\usage{
logit(p)
}
\arguments{
\item{p}{Number that you wish to transform}
}
\description{
This function takes a number and transforms it to
the logit scale
}
\examples{
logit(p)
}
