% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchResolution.R
\name{matchResolution}
\alias{matchResolution}
\title{Resample a raster from a reference}
\usage{
matchResolution(x, ref, method = "bilinear", filename = "", ...)
}
\arguments{
\item{x}{\code{Raster*} object to resample}

\item{ref}{Reference \code{Raster*} object with parameters that \code{x} should be resampled to.}

\item{method}{Character. Method used to compute values for the resampled raster. Can be \code{'bilinear'} for bilinear
interpolation or \code{'ngb'} for nearest neighbor interpolation.}

\item{filename}{Character. Output filename without directory. Can contain file extension (supported by writeRaster)
is in memory, will be written in working directory}

\item{...}{Other arguments passed to \code{\link[raster]{writeRaster}}}
}
\value{
A \code{Raster*} object
}
\description{
Resample a raster from a reference
}
\details{
\code{x} and \code{ref} must have defined CRS. You can assign a CRS using \code{\link[raster]{crs}}. If the CRS don't match, \code{x} is projected to \code{ref} CRS. \code{x} keeps its original extent, even if \code{ref} extent is larger.
}
