% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_caryAnnual.R
\name{met_caryAnnual}
\alias{met_caryAnnual}
\title{Annual mean monthly RH, Temp, and Precip}
\usage{
met_caryAnnual(start, end, path = "", center = TRUE)
}
\arguments{
\item{start}{start year}

\item{end}{end year}

\item{path}{file path to csv if not in working directory}

\item{center}{Should that data be centered? default is TRUE}
}
\description{
Function to create Annual Met from Cary weather stations
}
\examples{
met_caryAnnual(1995, 2005)
}
