% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_caryMonthly.R
\name{met_caryMonthly}
\alias{met_caryMonthly}
\title{Monthly mean monthly RH, Temp, and Precip}
\usage{
met_caryMonthly(start, end, path = "", center = TRUE)
}
\arguments{
\item{start}{start month (yyyymm)}

\item{end}{end month (yyyymm)}

\item{path}{file path to csv if not in working directory}

\item{center}{Should that data be centered? default is TRUE}
}
\description{
Function to create Monthly Met from Cary weather stations
}
\examples{
met_caryAnnual(199505, 200511)
}
