% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mna_tick.R
\name{mna_tick}
\alias{mna_tick}
\title{Mouse MNA closest to tick sampling}
\usage{
mna_tick(tick.date, mouse.date, known.states)
}
\arguments{
\item{tick.date}{vector of dates from tick sampling}

\item{mouse.date}{vecor of dates from mouse captures}

\item{known.states}{known states matrix}
}
\description{
Function that finds the minimum number alive of mice nearest in time to
tick sampling date, returns a vector of MNA
}
