% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Split data into training and testing sets}
\usage{
partition(x, type = "group holdout", p = 0.75, groups = min(5,
  length(x)), returnTrain = TRUE)
}
\arguments{
\item{x}{A vector used for splitting data}

\item{type}{Character. How should data be split? Valid values are "random holdout" , "group holdout" or "kfold"}

\item{p}{percentage of data that goes to training set (holdout) or to each fold (1/k)}

\item{groups}{For "group holdout" and when x is numeric, this is the number of breaks in the quantiles}

\item{returnTrain}{Logical indicating whether training data or testing data should be returned}
}
\description{
Returns the row indices of \code{x} that should go to training or testing.
}
\details{
Three types of splits are currently implemented. "random holdout" randomly select \code{p} percents of \code{x} for the training set. 'group holdout" first groups \code{x} into \code{groups} quantiles and randomly samples within them. "kfold" creates k folds where p percent of the data is used for training in each fold. This function is a wrapper around two functions of \code{caret} package: \code{\link[caret]{createDataPartition}} and \code{\link[caret]{createDataPartition}}
}
\seealso{
\code{\link[caret]{createDataPartition}}
}
