% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spdf2df}
\alias{spdf2df}
\title{Convert a SPDF to data.frame}
\usage{
spdf2df(x, xy = F)
}
\arguments{
\item{x}{A \code{\link[sp]{SpatialPoints}} object}

\item{xy}{Logical. If TRUE, coordinates of \code{x} are added to data.frame}
}
\description{
Convert a \code{\link[sp]{SpatialPoints}} to data.frame. Coordinates of \code{x} can be added to the data.frame using \code{xy = T}
}
