% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_type.R
\name{surv_type}
\alias{surv_type}
\title{Tick Survival Data Intake function}
\usage{
surv_type(file, subset, season = NULL)
}
\arguments{
\item{file}{Path to CSV}

\item{subset}{one of "Fed Larvae", "Flat Larvae", "Fed Nymph", "Flat Nymph", "Overwintering Nymph"}

\item{season}{default = NULL, must set to "winter or "summer for larvae}
}
\description{
This funciton reads the raw csv file containing tick survival data
from the Cary institute survival experiments (soil cores) and transforms
the data to be fed to JAGS
}
\examples{
surv.type(file, "Fed Nymph")
}
