% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theilSen}
\alias{theilSen}
\title{Theil-Sen slope}
\usage{
theilSen(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
numeric; Theil-Sen slope
}
\description{
Calculate the Theil-Sen slope from a time series. This is a wrapper around \code{\link[trend]{sens.slope}}
}
\examples{
x <- rnorm(100)
theilSen(x)
}
\seealso{
\code{\link[trend]{sens.slope}}
}
