% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_cary.R
\name{tick_cary}
\alias{tick_cary}
\title{Tick Data Intake}
\usage{
tick_cary(grid, type, path = "")
}
\arguments{
\item{grid}{What grid do you want data? One of Green Control, Henry Control, Tea Control, Green Experimental, Henry Experimental, Tea Experimental}

\item{type}{How should the population counted? One of density, individual, or delta}

\item{path}{File path to csv, default = ""}
}
\description{
Raw tick data was not great, so the following function just takes the cleaned data
and subsets for a specified grid and data type
}
\examples{
tick_cary(grid = "Green Control", type = "individual")
}
