% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFoster.R
\name{trainFoster}
\alias{trainFoster}
\title{Trains a kNN model from reponse variables (Y) and predictors (X) and return the accuracy of the model.
1. Divides the data in training and testing sets
2. If more than one k is given in input, performs parameter tuning using k-fold resampling
3. Picks the optimum value of k and train a model on all the training data
4. Predict values of the testing set
5. Compute accuracy metrics.}
\usage{
trainFoster(x, y, method = "kNN", k = 1, type = "randomForest",
  ntree = 200, mtry = NULL, fracTrain = 0.75, nFolds = 5)
}
\arguments{
\item{x}{A dataframe of predictors variables X for reference observations.
row names of X are identification of observations
X cannot contain missing values}

\item{y}{A matrix or dataframe of response variables Y for the reference observations.
row names of Y are identification of reference observations}

\item{method}{Character. Type of model used for the imputation. Currently supports k-NN only}

\item{k}{Numeric. Number of nearest neighbors}

\item{type}{Character. Type of distance metric ('euclidean','randomForest' etc.) if method='kNN'}

\item{fracTrain}{Numeric. Fraction of observations used for training. Remaining observations go to the testing set}

\item{nFolds}{Numeric. During parameter tuning, k-fold resmapling is performed.}

\item{...}{Additional arguments to control kNN method and type (e.g. ntree, mtry)}
}
\value{
predicted Y variables at targets
}
\description{
Trains a kNN model from reponse variables (Y) and predictors (X) and return the accuracy of the model.
1. Divides the data in training and testing sets
2. If more than one k is given in input, performs parameter tuning using k-fold resampling
3. Picks the optimum value of k and train a model on all the training data
4. Predict values of the testing set
5. Compute accuracy metrics.
}
