% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varCor.R
\name{varCor}
\alias{varCor}
\title{Calculate correlation between variables and discard the most correlated ones}
\usage{
varCor(x, vars = 1:ncol(x), use = "everything", method = c("pearson",
  "kendall", "spearman"), cutoff = 0.9, verbose = FALSE,
  names = FALSE, exact = ncol(x) < 100)
}
\arguments{
\item{x}{a numeric vector, matrix or dataframe}

\item{vars}{Numeric or character vector indicatig which variables (columns) should be considered. By default, all columns are used.}

\item{use}{an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"}, \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated.}

\item{cutoff}{Numeric value for the pair-wise aboslute correlation cutoff}

\item{verbose}{logical. Should details be printed?}

\item{names}{logical. Should column names (TRUE) pr column index (FALSE) be returned?}

\item{exact}{logical. Should the average correlations be recomputed at each step? See details in documentation of \code{\link[caret]{findCorrelation}}}
}
\value{
A list containing a new dataframe without most correlated variables and a vector with the columns index or names that have been removed
}
\description{
This is a wrapper function of \code{\link[caret]{findCorrelation}}.
}
