% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{varImp}
\alias{varImp}
\title{Returns variable importance}
\usage{
varImp(model, scaled = TRUE)
}
\arguments{
\item{model}{A yai object}

\item{scaled}{Should impotance values be centered and scaled}
}
\description{
When RF is used to find nearest neighbours, the importance of each variable in the RF trees is calculated
}
