% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndices.R
\name{calcIndices}
\alias{calcIndices}
\title{calcIndices}
\usage{
calcIndices(x, method = "TC", filename = "", ...)
}
\arguments{
\item{x}{Composite raster composed of the spectral bands used to calculate the indices. X can be a path to raster or RasterBrick object}

\item{method}{Which method is used to calculate the indices. Default is Tasseled Cap. FOr a list of other methods see RStoolbox::spectralIndices}

\item{filename}{character. Optional output filename. If no filename is provided, the output is written to disk only if they cannot be stored in RAM (raster::canProcessinMemory)}

\item{...}{Further arguments passed to RStoolbox:tasseledCap, RStoolbox:spectralIndices or raster::writeRaster}
}
\value{
A Raster object with spectral indices
}
\description{
Calculates spectral indices from multibands spectral images
}
