% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defaultTemporalSummary}
\alias{defaultTemporalSummary}
\title{Default temporal summary}
\usage{
defaultTemporalSummary(x)
}
\arguments{
\item{x}{Vector of numeric values}
}
\value{
Named vector with median, IQR and slope
}
\description{
Calculates median, IQR and Theil Sen slope (\code{\link[trend]{sens.slope}}).
This function is usually called within \code{\link[foster]{temporalMetrics}}
}
\examples{
x <- rnorm(100)
defaultTemporalSummary(x)
}
\seealso{
\code{\link[foster]{temporalMetrics}}, \code{\link[trend]{sens.slope}}
}
