% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\title{Remove cells in the neighbourhood of a boundary cell}
\usage{
edges(x, w, filename = "", ...)
}
\arguments{
\item{x}{A \code{Raster* object}}

\item{w}{Numeric. Size of the window around each cell. Must be an odd number.}

\item{filename}{Character (optional). If the output is written to disk, full path to location including filename}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}}
}
\description{
Assigns NA value to all cells having a NA values within their \code{wxw} neighbourhood.
}
