% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNN.R
\name{findNN}
\alias{findNN}
\title{Train a kNN (or regression model) from reponse variables (Y) and predictors (X) at reference location}
\usage{
findNN(x, y, inTrain = NULL, inVal = NULL, k = 1,
  method = "randomForest", ntree = 200, mtry = NULL)
}
\arguments{
\item{x}{A matrix or dataframe of predictors variables X for reference observations.Row names of X are identification of observations. X cannot contain missing values}

\item{y}{A matrix or dataframe of response variables Y for the reference observations. Row names of Y are identification of reference observations}

\item{inTrain}{Optional numeric vector indicating which rows of x and y go to training.}

\item{inVal}{Optional numeric vector indicating which rows of x and y go to training. If left empy, all row that are not in inTrain are used for validation}

\item{k}{Integer. Number of nearest neighbors}

\item{method}{Character. Which nearness netrics is used to computed the nearest neighbors. Default is \code{"randomForest"}. Other methods are listed in \code{\link[yaImpute]{yai}}}

\item{ntree}{Number of classification or regression trees drawn for each response variable. Default is 200}

\item{mtry}{Number of X varaibles picked randomly to split each node. Default is sqrt(number of X variables)}
}
\value{
predicted Y variables at targets
}
\description{
Train a kNN (or regression model) from reponse variables (Y) and predictors (X) at reference location
}
