% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalMultiBand.R
\name{focalMultiBand}
\alias{focalMultiBand}
\title{Apply a spatial filter to a Raster object}
\usage{
focalMultiBand(x, w, fun, filename = "", na.rm = TRUE, pad = FALSE,
  padValue = NA, NAonly = FALSE, keepNA = T, ...)
}
\arguments{
\item{x}{A Raster* object}

\item{w}{Matrix of weights (moving window). A 3x3 windows with weights of 1 would be \code{w=matrix(1,nr=3,nc=3)} for example.}

\item{fun}{function (optional). The function should accept a vector of values and return a single number (e.g. mean). It should also accept a \code{na.rm} argument.}

\item{filename}{Character (optional). If the output is written to disk, full path to location including filename}

\item{na.rm}{logical. If TRUE (default), NA are removed from computation}

\item{pad}{logical. IF TRUE, rows and columns are added around \code{x} to avoid removing border cells.}

\item{padValue}{numeric. Value of \code{pad} cells. Usually set to NA and used in combination with \code{na.rm=TRUE}}

\item{NAonly}{logical. If TRUE only cell values that are NA are replaced with the computed focal values}

\item{keepNA}{Logical. If TRUE (default), NA cells of \code{x} are unchanged}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}}
}
\description{
Apply a spatial filter to a RatsterLayer or all layers of a RatserStack or RasterBrick object. The mathematical operation applied within the neighbourhood can be done by using a function (\code{fun}) or by adjusting the weights of the matrix \code{w}.
}
\details{
If the Raster contains NA values, applying a function \code{fun} with \code{na.rm=TRUE} or adjusting the weights of \code{w} are not equivalent (in that case the result would be wrong). During filtering, NA cells of \code{x} can get a value assigned if \code{na.rm=T}. In order to keep these cells value to NA, the argument \code{keepNA=TRUE} can be used (default).
}
