% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleValues.R
\name{getSampleValues}
\alias{getSampleValues}
\title{Get raster cell values at samples}
\usage{
getSampleValues(x, s)
}
\arguments{
\item{x}{A raster layer}

\item{s}{Locations of the samples. Object of class \code{SpatialPointsDataFrame}, usually generated with \code{getSampleXY}}
}
\value{
A \code{SpatialPointsDataFrame} object containing pixel values corresponding to sample locations
}
\description{
Get raster cell values at samples
}
