% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{GetSampleValues}
\alias{GetSampleValues}
\title{Get pixel values for sample locations}
\usage{
GetSampleValues(x, s)
}
\arguments{
\item{x}{A raster layer}

\item{s}{Locations of the samples. Object of class \code{SpatialPointsDataFrame}. Can be generated with \code{getSampleXY}}
}
\value{
A \code{SpatialPointsDataFrame} object containing pixel values corresponding to sample locations
}
\description{
Get pixel values for sample locations
}
