% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleValues.R
\name{getSampleValues}
\alias{getSampleValues}
\title{Get raster cell values at sample points}
\usage{
getSampleValues(x, s, append = FALSE, filename = "", ...)
}
\arguments{
\item{x}{A raster layer}

\item{s}{Location of the sample points. Object of class \code{SpatialPointsDataFrame}, usually generated with \code{\link[foster]{getSample}}}

\item{append}{Should the values be appended to s? Default is FALSE}

\item{filename}{Character. Output filename including path to directory. File will be automatically saved as an ESRI Shapefile and any extension in \code{filename} will be overwritten}

\item{...}{Additional arguments passed to \code{\link[rgdal]{writeOGR}}}
}
\value{
A \code{\link[sp]{SpatialPoints}} object containing cells values corresponding to samples
}
\description{
Given a raster layer and a SpatialPointsDataFrame object, the functions returns a SpatialPointsDataFrame objects with the values of the raster at sample points.
}
\seealso{
\code{\link[raster]{extract}}
}
