% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleXY.R
\name{getSampleXY}
\alias{getSampleXY}
\title{Stratified random sampling}
\usage{
getSampleXY(x, layers = seq(1, raster::nlayers(x), 1), n, mindist = 0,
  nClasses = 5, ...)
}
\arguments{
\item{x}{A \code{Raster*} object used to generate random sample}

\item{layers}{Vector containing the band numbers of \code{x} used in stratification. By default, all layers of x are used}

\item{n}{Sample size}

\item{mindist}{Minimum distance between samples (in units of \code{x}). Default is 0.}

\item{nClasses}{Number of strata. Default is 5.}

\item{...}{Further arguments passed to \code{\link[RStoolbox]{unsuperClass}} or \code{\link[raster]{writeRaster}} to control the kmeans algorithm or writing parameters}
}
\value{
A \code{SpatialPointsDataFrame} object containing sample locations
}
\description{
Generate a minimum distance stratified random sample
}
\details{
\code{x} is stratified using kmean clustering from \code{\link[RStoolbox]{unsuperClass}}. Default kmean clustering method is Hartigan-Wong algorithm. The algorithm might not converge and output "Quick Tansfer" warning. If this is the case, we suggest decreasing \code{nClasses}. Also, if \code{minDist} is too large, the algorithm might not be able to select enough sample per strata. In that case, the warning "Exeeded maximum number of runs for strata" is displayed.
}
