% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleXY.R
\name{getSampleXY}
\alias{getSampleXY}
\title{Generate a minimum distance stratified random sample}
\usage{
getSampleXY(x, layers = seq(1, nlayers(x), 1), n, mindist = 500,
  nClasses = 10, ...)
}
\arguments{
\item{x}{A \code{Raster*} object used to generate random sample}

\item{layers}{Vector containing the layer bands used in stratification. By defualt all lyers of x are used}

\item{n}{Sample size}

\item{mindist}{Minimum distance between samples (in meters)}

\item{nClasses}{Number of strata. Default is 10}

\item{...}{Further arguments passed to \code{RStoolbox::unsuperClass}, to control the kmeans algorithm}
}
\value{
A \code{SpatialPointsDataFrame} object containing sample locations
}
\description{
Generate a minimum distance stratified random sample
}
