% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeFoster.R
\name{imputeFoster}
\alias{imputeFoster}
\title{Returns Y values at targets}
\usage{
imputeFoster(model = NULL, X.trgs = NULL, ...)
}
\arguments{
\item{model}{A model trained at reference locations, ususally created from \code{modelFoster}}

\item{Xtrgs}{Predictors (X) values at targets}
}
\value{
predicted Y variables at targets
}
\description{
Returns Y values at targets
}
