% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeFoster.R, R/predictTrgs.R
\name{imputeFoster}
\alias{imputeFoster}
\title{Returns Y values at targets}
\usage{
imputeFoster(model = NULL, X.trgs = NULL, filename = "", par = F,
  n.cores = 1, ...)

imputeFoster(model = NULL, X.trgs = NULL, filename = "", par = F,
  n.cores = 1, ...)
}
\arguments{
\item{model}{A model trained at reference locations, ususally created from \code{modelFoster}}

\item{filename}{Character. Path to file written to disk (may include the extension)}

\item{par}{Logical. Should imputation be performed on parallel threads?}

\item{n.cores}{Integer. Number of parallel threads (relevant only if par=TRUE)}

\item{Xtrgs}{Predictors (X) values at targets}

\item{model}{A model trained at reference locations, ususally created from \code{modelFoster}}

\item{Xtrgs}{Predictors (X) values at targets}

\item{filename}{Character. Path to file written to disk (may include the extension)}

\item{par}{Logical. Should imputation be performed on parallel threads?}

\item{n.cores}{Integer. Number of parallel threads (relevant only if par=TRUE)}
}
\value{
predicted Y variables at targets

predicted Y variables at targets
}
\description{
Returns Y values at targets

Returns Y values at targets
}
