% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchExtent.R
\name{MatchExtent}
\alias{MatchExtent}
\title{Crop and mask input raster to match the extent of the reference}
\usage{
MatchExtent(x, ref, output_dir = NA, output_name = NA,
  output_format = "tif", overwrite = TRUE)
}
\arguments{
\item{x}{Data to crop and mask. May be a \code{raster} object or a path to a raster file.}

\item{ref}{Reference raster. May be a \code{raster} object or a path to a raster file.}

\item{output_dir}{Path to the output directory. If not provided, output is saved in the same directory as \code{x}.}

\item{output_name}{Filename of the resampled raster without extension. If not provided \code{x} name is used with an appendix (_resampled)}

\item{output_format}{Extension of the resampled raster. (Default=tif).}

\item{overwrite}{Logical. Should the output raster be overwritten?}
}
\description{
Crop and mask input raster to match the extent of the reference
}
