% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchExtent.R
\name{matchExtent}
\alias{matchExtent}
\title{Match the extent of a reference Raster}
\usage{
matchExtent(x, ref, mask = F, inverse = FALSE, maskValue = NA,
  filename = "", ...)
}
\arguments{
\item{x}{Raster* object}

\item{ref}{Raster* object with extent that \code{x} shoudl be cropped to}

\item{mask}{logical. Should x be masked by \code{ref} cells that have the value \code{maskValue}}

\item{inverse}{logical. If TRUE, cells of \code{ref} that are not \code{maskvalue} are masked}

\item{maskValue}{Value of \code{ref} cells that should be masked in \code{x}}

\item{filename}{Character. Output filename}

\item{...}{Other arguments passed to \code{writeRaster}}
}
\value{
A \code{Raster*} object
}
\description{
Crops or extends the extent of a raster to the extent of a reference. Some cells of the reference raster can optionally be masked based on their values.
}
\details{
\code{x} and \code{ref} need to have the same CRS, spatial resolution and origin. If this is not the case, you can use \code{\link[foster]{MatchResolution}}.
}
