% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchExtent.R
\name{matchExtent}
\alias{matchExtent}
\title{Match the extent of a reference Raster}
\usage{
matchExtent(x, ref, mask = F, inverse = FALSE, maskValue = NA,
  updatevalue = NA, updateNA = FALSE, filename = "", ...)
}
\arguments{
\item{x}{Raster* object}

\item{ref}{Raster* object with extent that \code{x} shoudl be cropped to}

\item{mask}{logical. Should x be masked by \code{ref} cells that have the value \code{maskValue}}

\item{maskValue}{logical. Value of \code{ref} cells that should be masked in \code{x}}

\item{filename}{Character. Output filename}

\item{...}{Other arguments passed to \code{writeRaster}}
}
\value{
A \code{Raster*} object
}
\description{
Crop and mask input raster to match the extent of the reference
}
\details{
\code{x} and \code{ref} need to have the same CRS, spatial resolution and origin. If this is not the case, you can use \code{\link[foster]{matchResolution}}. For more information about cropping and masking function you can refer to \code{\link[raster]{crop}} and \code{\link[raster]{mask}}
}
