% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchResolution.R
\name{MatchResolution}
\alias{MatchResolution}
\title{Resample lidar-based rasters to match satellite images}
\usage{
MatchResolution(x, ref, output_dir = NA, output_name = NA,
  output_format = "tif", overwrite = TRUE)
}
\arguments{
\item{x}{Data to resample. May be an existing \code{raster} object or a path to a raster file.}

\item{ref}{Reference raster (e.g. wall-to-wall satellite image). \code{x} will be resampled to match \code{ref}. May be an existing \code{raster} object or a path to a raster file.}

\item{output_dir}{Path to the output directory. If not provided, output is saved in the same directory as \code{x}.}

\item{output_name}{Filename of the resampled raster without extension. If not provided \code{x} name is used with an appendix (_resampled)}

\item{output_format}{Extension of the resampled raster. (Default=tif).}

\item{overwrite}{Logical. Should the output raster be overwritten? (Default=TRUE)}
}
\description{
Resample lidar-based rasters to match satellite images
}
