% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchResolution.R
\name{matchResolution}
\alias{matchResolution}
\title{Match the resolution of two Raster objects}
\usage{
matchResolution(x, ref, method = "bilinear", filename = "", ...)
}
\arguments{
\item{x}{\code{Raster*} object to resample}

\item{ref}{Reference \code{Raster*} object with parameters that \code{x} should be resampled to.}

\item{method}{Character. Method used to compute values for the resampled raster. Can be \code{'bilinear'} for bilinear
interpolation or \code{'ngb'} for nearest neighbor interpolation.}

\item{filename}{Character. Output filename (optional)}

\item{...}{Other arguments passed to \code{\link[raster]{writeRaster}}}
}
\value{
A \code{Raster*} object
}
\description{
Successsively projects and resamples a Raster coordinate system and spatial resolution to the reference
}
\details{
\code{x} and \code{ref} must have defined CRS (can be assigned using \code{\link[raster]{projection}}). If the CRS don't match, \code{x} is projected to \code{ref} CRS prior to resampling. \code{x} doesn't inherits the extent of \code{ref}.
}
