% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFoster.R
\name{scatter}
\alias{scatter}
\title{scatterplot with optional information on the errors between x and y.}
\usage{
scatter(predicted, observed, by = NULL, axisorder = "OP",
  xlab = "Predicted", ylab = "Observed", title = NULL, info = T,
  size.point = 4, shape.point = 1, size.info = 6, position = 0,
  positionauto = T, lowerlimit = NA, upperlimit = NA, alpha = 1,
  add.reg.line = F, rug = F, label = c("count", "R2", "bias",
  "bias_per", "RMSE", "RMSE_per", "slope", "intercept"),
  label_text = c("n", "R2", "bias", "bias\%", "RMSE", "RMSE\%", "slope",
  "intercept"))
}
\arguments{
\item{predicted}{a vector of predicted data.}

\item{observed}{a vector of observed data.}

\item{by}{optional grouping variable. Can be character or factor}

\item{axisorder}{Optional. Set to \code{PO} (predicted-observed) to plot predicted (\code{y}) on the y-axis (this is the default). Set to \code{OP} (observed-predicted) to plot observed (\code{x}) on the y-axis.}

\item{xlab}{Optional. Title of the x-axis}

\item{ylab}{Optional. Title of the y-axis}

\item{info}{A logical value indictating whether information on count, bias and RMSE should be added to the plot.}

\item{position}{Determines the position of the info box}

\item{positionauto}{A logical value indicating whether the position of the info box should be optimized automaticaly}

\item{lowerlimit}{A value determining the lower limit of the x and y axis}

\item{upperlimit}{A value determining the upper limit of the x and y axis}

\item{alpha}{Define the transparency of the points. 0 - fully transparent, 1 - opaque.}

\item{add.reg.line}{Logical. Should the regression line be added to the plot? Regression coeficients are calculated automatically.}

\item{rug}{Logical. Add marginal rug to the plot.}

\item{label}{A character vector containing the name of the metrics that are annotated on the plot. Must be in c("count","R2","bias","bias_per","RMSE","RMSE_per","slope","intercept")}

\item{label_text}{A character vector containing of the same length as label contaning th name of each label. Default is c("n","R2","bias","bias\%","RMSE","RMSE\%","slope","intercept")}
}
\value{
a scatterplot of \code{x} and \code{y}.
}
\description{
scatterplot with optional information on the errors between x and y.
}
