% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdf2df.R
\name{spdf2df}
\alias{spdf2df}
\title{Convert a SPDF to data.frame}
\usage{
spdf2df(x, xy = F)
}
\arguments{
\item{x}{A \code{SpatailPointsDataFrame} object}

\item{xy}{Logical. If TRUE, coordinates of \code{x} are added to data.frame}
}
\description{
Convert a SpatialPointsDataFrame to data.frame. Coordinates of \code{x} can be added to the data.frame by using \code{xy=T}
}
