% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalMetrics.R
\name{temporalMetrics}
\alias{temporalMetrics}
\title{Calculate temporal metrics}
\usage{
temporalMetrics(s, metrics = "defaultTemporalSummary", prefix = NULL,
  filename = "", ...)
}
\arguments{
\item{x}{Input \code{raster::stack} object containing a time series of spectral indices (e.g. NDVI). Input stack may be generated with \code{calcIndices}}

\item{fun}{Name of a function (e.g. \code{"median"} or \code{"myFunction"}) used to process the time series data. Function should return a named \code{list}. \code{default_temporal_summary} function used as the default returns median, IQR, and slope values.}

\item{output_dir}{Path to the output directory. If not provided, output is saved in the same directory as \code{x}.}

\item{output_name}{Filename of the resampled raster without extension. If not provided \code{x} name is used with an appendix (_resampled)}

\item{output_format}{Extension of the resampled raster. (Default=tif).}

\item{overwrite}{Logical. Should the output raster be overwritten?}
}
\description{
Calculate temporal metrics
}
