% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile}
\alias{tile}
\title{Split a raster into tiles}
\usage{
tile(x, nx, ny, filename = "", suffix = NULL, ...)
}
\arguments{
\item{x}{Raster object to split}

\item{nx}{Number of horizontal cells in the splitting grid}

\item{ny}{Number of vertical cells in the splitting grid}

\item{filename}{Character (optional). Output filename including path to
directory and eventually extension}

\item{suffix}{Character appended to filename to differentiate tiles
(must have length nx x ny). If left NULL, tiles will be numbered by columns
and rows}

\item{...}{Additional parameters passed to \code{\link[raster]{writeRaster}}}
}
\description{
This function is used to split a raster into smaller tiles. The raster is
split in a grid pattern with nx columns and ny rows.
}
\seealso{
\code{\link[raster]{crop}}
}
